Refactor:리팩토링 페이지에 따라 다른정보 보여주기
