+++ constant.ts
@@ -5,7 +5,7 @@ const PAGE_CONTENT = {
     button: '로그인',
     text: '회원이 아니신가요?',
     path: '/signin',
-    link: '로그인 하기',
+    linkText: '로그인 하기',
     email: 'signin-email',
     password: 'signin-password'
   },
@@ -13,10 +13,22 @@ const PAGE_CONTENT = {
     button: '회원가입',
     text: '이미 회원이신가요?',
     path: '/signup',
-    link: '회원가입 하기',
+    linkText: '회원가입 하기',
     email: 'signup-email',
     password: 'signup-password'
   },
 };
 
-export { BASE_URL, PAGE_CONTENT }
+const BUTTON = [
+  {
+    id: 'linkRemove',
+    name: '삭제하기',
+  },
+  {
+    id: 'folderAdd',
+    name: '폴더에 추가',
+  },
+];
+
+
+export { BASE_URL, PAGE_CONTENT, BUTTON }
